mcp-vocabulary/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
prometheus-client>=0.20.0
python-json-logger==2.0.7
sentence-transformers==3.3.1
asyncpg==0.30.0
pyyaml>=6.0
